improve style for desktop
